Add first pass at ADA integration
We can get two types of data back from ADA - Classes or ProgrammeItems.
For the work on the Radio brand page we only care about Classes. There
is a AdaClassService because eventually there will be an
AdaProgrammeItemService (or something named similar).

Tests for the mapper and domain model are in there, but have not yet
done a test for the service yet.